+++ src/FractalServiceProvider.php
@@ -2,9 +2,7 @@
 
 namespace Spatie\Fractal;
 
-use Spatie\Fractalistic\Fractal;
 use Illuminate\Support\ServiceProvider;
-use League\Fractal\Serializer\SerializerAbstract;
 use Spatie\Fractal\Console\Commands\TransformerMakeCommand;
 
 class FractalServiceProvider extends ServiceProvider
@@ -35,24 +33,5 @@ class FractalServiceProvider extends ServiceProvider
         $this->app->bind('laravel-fractal', function (...$arguments) {
             return fractal(...$arguments);
         });
-
-        $this->app->alias(Fractal::class, 'laravel-fractal');
-    }
-
-    /**
-     * Set the default serializer.
-     *
-     * @param \Spatie\Fractal\Fractal                              $fractal
-     * @param string|\League\Fractal\Serializer\SerializerAbstract $serializer
-     *
-     * @return \Spatie\Fractal\Fractal
-     */
-    protected function setDefaultSerializer($fractal, $serializer)
-    {
-        if ($serializer instanceof SerializerAbstract) {
-            return $fractal->serializeWith($serializer);
-        }
-
-        return $fractal->serializeWith(new $serializer());
     }
 }