+++ src/helpers.php
@@ -1,7 +1,6 @@
 <?php
 
-use Spatie\Fractalistic\Fractal;
-use League\Fractal\Serializer\SerializerAbstract;
+use Spatie\Fractal\Fractal;
 
 if (! function_exists('fractal')) {
     /**
@@ -9,24 +8,10 @@ if (! function_exists('fractal')) {
      * @param null|callable|\League\Fractal\TransformerAbstract $transformer
      * @param null|\League\Fractal\Serializer\SerializerAbstract $serializer
      *
-     * @return \Spatie\Fractalistic\Fractal
+     * @return \Spatie\Fractal\Fractal
      */
     function fractal($data = null, $transformer = null, $serializer = null)
     {
-        $fractal = Fractal::create($data, $transformer, $serializer);
-
-        $serializer = config('laravel-fractal.default_serializer');
-
-        if (! empty($serializer)) {
-            if ($serializer instanceof SerializerAbstract) {
-                $fractal->serializeWith($serializer);
-            } elseif ($serializer instanceof Closure) {
-                $fractal->serializeWith($serializer());
-            } else {
-                $fractal->serializeWith(new $serializer());
-            }
-        }
-
-        return $fractal;
+        return Fractal::create($data, $transformer, $serializer);
     }
 }